Seperate in different modules
