+++ src/utils/Tooling.ts
@@ -1,7 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-// uint8array to string
+import * as ccfapp from "@microsoft/ccf-app";
+import { IKeyReleasePolicyProps } from "../policies/IKeyReleasePolicyProps";
+import { ccf } from "@microsoft/ccf-app/global";
+
+/**
+ * Converts a Uint8Array to a string representation.
+ * @param uInt8array - The Uint8Array to convert.
+ * @returns The string representation of the Uint8Array.
+ */
 export const convertUint8ArrayToString = (uInt8array: Uint8Array): string => {
   let stringRepresentation = "";
   for (let i = 0; i < uInt8array.length; i++) {
@@ -10,9 +18,87 @@ export const convertUint8ArrayToString = (uInt8array: Uint8Array): string => {
   return stringRepresentation;
 };
 
-// ArrayBuffer to hex
+/**
+ * Converts an ArrayBuffer to a hexadecimal string representation.
+ * @param buf - The ArrayBuffer to convert.
+ * @returns The hexadecimal string representation of the ArrayBuffer.
+ */
 export const arrayBufferToHex = (buf: ArrayBuffer) => {
   return Array.from(new Uint8Array(buf))
     .map((n) => n.toString(16).padStart(2, "0"))
     .join("");
+};
+
+/**
+ * Parses query parameters from a request.
+ * @param request - The request object containing the query parameters.
+ * @returns An object representing the parsed query parameters.
+ */
+export const queryParams = (request: ccfapp.Request) => {
+  const elements = request.query.split("&");
+  let obj = {};
+  for (let inx = 0; inx < elements.length; inx++) {
+    const param = elements[inx].split("=");
+    obj[param[0]] = param[1];
+    console.log(`Query: ${param[0]} = ${param[1]}`);
+  }
+  return obj;
+};
+
+/**
+ * Checks if a given string represents a PEM public key.
+ * @param key - The PEM key to be checked.
+ * @returns A boolean indicating whether the string is a PEM public key.
+ */
+export const isPemPublicKey = (key: string): boolean => {
+  const beginPatternLiteral = /-----BEGIN PUBLIC KEY-----\\n/;
+  const endPatternLiteral = /\\n-----END PUBLIC KEY-----\\n$/;
+  const beginPatternNewline = /-----BEGIN PUBLIC KEY-----\n/;
+  const endPatternNewline = /\n-----END PUBLIC KEY-----\n$/;
+  const contentPattern = /([\s\S]+)/;
+
+  const isLiteralNewline =
+    beginPatternLiteral.test(key) && endPatternLiteral.test(key);
+  const isNewline =
+    beginPatternNewline.test(key) && endPatternNewline.test(key);
+
+  console.log("isLiteralNewline:", isLiteralNewline);
+  console.log("isNewline:", isNewline);
+
+  return isLiteralNewline || isNewline;
+};
+
+/**
+ * Retrieves the key release policy from a key release policy map.
+ * @param keyReleasePolicyMap - The key release policy map.
+ * @returns The key release policy as an object.
+ */
+export const getKeyReleasePolicy = (
+  keyReleasePolicyMap: ccfapp.KvMap,
+): IKeyReleasePolicyProps => {
+  const result: IKeyReleasePolicyProps = {};
+  keyReleasePolicyMap.forEach((values: ArrayBuffer, key: ArrayBuffer) => {
+    const kvKey = ccf.bufToStr(key);
+    const kvValue = JSON.parse(ccf.bufToStr(values));
+    result[kvKey] = kvValue;
+    console.log(`key policy item with key: ${kvKey} and value: ${kvValue}`);
+  });
+  console.log(
+    `Resulting key release policy: ${JSON.stringify(
+      result,
+    )}, keys: ${Object.keys(result)}, keys: ${Object.keys(result).length}`,
+  );
+  return result;
+};
+
+/**
+ * Sets the key headers.
+ * @returns An object containing the key headers.
+ */
+export const setKeyHeaders = (): { [key: string]: string } => {
+  const headers: { [key: string]: string } = {
+    "cache-control": "max-age=254838",
+    date: new Date().toUTCString(),
+  };
+  return headers;
 };
