+++ test/e2e-test/src/api.ts
@@ -61,7 +61,7 @@ export default class Api {
       let statusCode = 0;
       request.on("data", (chunk: string | Buffer) => {
         if (responseType === "json") {
-          console.log('responsePromise chunk: ', chunk);
+          console.log("responsePromise chunk: ", chunk);
           data += chunk;
         } else {
           chunks.push(chunk as Buffer);
